+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Improvements:
 * Core: upgraded jQueryUI to 1.8.19
 * Core: SELECT 1 ... LIMIT 1 in favor of SELECT COUNT(*), more optimised queries.
 * Spoon: merged changes.
+* Core: create minify cache folders if they do not exist.
 
 
 3.3.4 (2012-04-17)